Question 6 faite au complet avec des modifications necessaires dans la question 7 ( modifier le type de la variable montant ( int >> double) dans les classes des operations.

+++ src/main/java/com/atoudeft/banque/OperationFacture.java
@@ -4,7 +4,7 @@ public class OperationFacture extends Operation{
     private String numFacture;
     private String desc;
 
-    public OperationFacture(TypeOperation typeOperation, int montant, String numFacture, String desc) {
+    public OperationFacture(TypeOperation typeOperation, double montant, String numFacture, String desc) {
         super(typeOperation, montant);
         this.numFacture = numFacture;
         this.desc = desc;
