+++ src/main/java/com/atoudeft/banque/OperationRetrait.java
@@ -2,7 +2,7 @@ package com.atoudeft.banque;
 
 public class OperationRetrait extends Operation{
 
-    public OperationRetrait(TypeOperation typeOperation, int montant) {
+    public OperationRetrait(TypeOperation typeOperation, double montant) {
         super(typeOperation, montant);
     }
 }
