+++ src/main/java/com/atoudeft/banque/OperationTransfer.java
@@ -3,7 +3,7 @@ package com.atoudeft.banque;
 public class OperationTransfer extends Operation{
     private String numCompteClient;
 
-    public OperationTransfer(TypeOperation typeOperation, int montant, String numCompteClient) {
+    public OperationTransfer(TypeOperation typeOperation, double montant, String numCompteClient) {
         super(typeOperation, montant);
         this.numCompteClient = numCompteClient;
     }
